renamed folders berfore shadcn install
