+++ src/components/shared/LeftSidebar/CommandPalette/CommandPalette.tsx
